requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
ecdsa==0.19.1
fastapi==0.115.11
h11==0.14.0
idna==3.10
inflection==0.5.1
packaging==24.2
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
